+++ src/data/colors.ts
@@ -1,3 +1,6 @@
+// In TypeScript 2.4 this can turn into an Enum
+// https://github.com/Microsoft/TypeScript/pull/15486
+
 export default {
   "gray-light": "#f8f8f8",
   "gray-regular": "#e5e5e5",