+++ src/lib/app_registry.tsx
@@ -3,6 +3,7 @@ import * as React from "react"
 import { AppRegistry, ViewProperties } from "react-native"
 import * as Relay from "react-relay"
 
+import Consignments from "./components/consignments"
 import LoadFailureView from "./components/load_failure_view"
 import Spinner from "./components/spinner"
 import Containers from "./containers/index"
@@ -99,6 +100,7 @@ class MyAccount extends RootContainer {
   }
 }
 
+AppRegistry.registerComponent("Consignments", () => Consignments)
 AppRegistry.registerComponent("Artist", () => Artist)
 AppRegistry.registerComponent("Home", () => Home)
 AppRegistry.registerComponent("Gene", () => Gene)