+++ src/lib/components/works_for_you/notification.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { Image, StyleSheet, TextStyle, TouchableWithoutFeedback, View, ViewProperties, ViewStyle } from "react-native"
+import { Image, StyleSheet, TextStyle, TouchableWithoutFeedback, View, ViewStyle } from "react-native"
 import * as Relay from "react-relay"
 
 import SwitchBoard from "../../native_modules/switch_board"
@@ -114,7 +114,7 @@ export default Relay.createContainer(Notification, {
         message
         artists
         artworks {
-          artists (shallow: true) {
+          artists(shallow: true) {
             href
           }
           ${(ArtworksGrid.getFragment("artworks"))}