+++ src/storiesRegistry.js
@@ -1,0 +1,7 @@
+
+import "./lib/components/text/__stories__/typography.story"
+import "./lib/components/buttons/__stories__/buttons.story"
+
+import "./lib/containers/__stories__"
+import "./lib/components/artist/__stories__"
+import "./lib/components/consignments/__stories__"