+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 ### Master
 
+#### Consignments
+
+- [dev] Adds component for artist search  - orta
+- [dev] Adds component for consignments todo  - orta
+- [dev] Adds a root component for the Consignments flow  - orta
+- [dev] Adds storybooks for consignments  - orta
+- Added the Welcome/Overview screen for consignments - orta
+
+### Emission
+
+- [dev] Allows toggling the back button by pressing space - orta
+
+### 1.3.6
+
+- Fixes missing compiled bundle - alloy
+
+### 1.3.5
+
+- Fixes pull-to-refresh scroll position bug on Home - sarah
+- Fixes an issue where the WFU component was fetching data for artworks itself rather than delegating that to the nested
+  artworks components - alloy
+
+### 1.3.4
+
+- Fixes loading state UI for artwork rail - luc
+
 ### 1.3.3
 
 - Ensured rails without artworks cannot be rendered - sarah
@@ -89,7 +115,7 @@
 ### 1.1.0-beta.4
 
 - GeneVC scrolls as expected, and can request data from Eigen - orta
-- Fix crash that would occur when tapping the ‘failed to load’ view retry button a second time - alloy
+- Fix crash that would occur when tapping the "failed to load" view retry button a second time - alloy
 - Make tapping the search bar open the search VC - sarah
 
 ### 1.1.0-beta.3
@@ -105,7 +131,7 @@
 - [dev] Turns on "keychain sharing" to fix a keychain bug in sim - orta
 - GeneVC now shows about information, and trending artists - orta
 - GeneVC now shows artworks for the gene - orta
-- Show ‘bid now’ call to action on artworks that are in an auction - alloy
+- Show "bid now" call to action on artworks that are in an auction - alloy
 
 ### 1.1.0-beta.2